# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sampler Book Library API",
        "description": "This is the documentation of the Sampler Book Library API, with this API you can check-in and out books after you register as an user",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/books/search": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Search books",
                "description": "Search by Title OR ISBN. If both are passed, ISBN will be used.",
                "operationId": "App\\Http\\Controllers\\BooksController::search",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Search books",
                                        "type": "string",
                                        "example": "Harry Potter"
                                    },
                                    "isbn": {
                                        "description": "Search books",
                                        "type": "string",
                                        "example": "1408855895"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mathing books.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Search books",
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "description": "Search books",
                                            "properties": {
                                                "isbn": {
                                                    "description": "Search books",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Must be a valid ISBN-10."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/api/books": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create book",
                "description": "Adds a new book to the library.",
                "operationId": "App\\Http\\Controllers\\BooksController::create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Create book",
                                        "type": "integer",
                                        "readOnly": "true",
                                        "example": "1"
                                    },
                                    "title": {
                                        "description": "Create book",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Harry Potter and the Philosopher's Stone"
                                    },
                                    "isbn": {
                                        "description": "Create book",
                                        "type": "string",
                                        "maxLength": 10,
                                        "example": "1408855895"
                                    },
                                    "published_at": {
                                        "description": "Create book",
                                        "type": "date",
                                        "example": "2014-01-22"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The book data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Create book",
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "description": "Create book",
                                            "properties": {
                                                "isbn": {
                                                    "description": "Create book",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Must be a valid ISBN-10."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/api/books/{bookId}/checkout": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Checkout book",
                "description": "Checks out a book.",
                "operationId": "App\\Http\\Controllers\\BooksController::checkout",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "ID of book",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book successfully checked out.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Checkout book",
                                            "type": "string",
                                            "example": "Book checkout successful, enjoy your reading!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Checkout book",
                                            "type": "string",
                                            "example": "Action could be completed."
                                        },
                                        "errors": {
                                            "description": "Checkout book",
                                            "properties": {
                                                "checked_out": {
                                                    "description": "Checkout book",
                                                    "type": "string",
                                                    "example": "Book is not available to checkout."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/api/books/{bookId}/checkin": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Checkin book",
                "description": "Checks in a book.",
                "operationId": "App\\Http\\Controllers\\BooksController::checkin",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "ID of book",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book successfully checked in.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Checkin book",
                                            "type": "string",
                                            "example": "Book checkin successful, combe back later!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Checkin book",
                                            "type": "string",
                                            "example": "Action could be completed."
                                        },
                                        "errors": {
                                            "description": "Checkin book",
                                            "properties": {
                                                "already_available": {
                                                    "description": "Checkin book",
                                                    "type": "string",
                                                    "example": "Book is already available."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Book": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Harry Potter and the Philosopher's Stone"
                    },
                    "isbn": {
                        "type": "string",
                        "maxLength": 10,
                        "example": "1408855895"
                    },
                    "published_at": {
                        "type": "date",
                        "example": "2014-01-22"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "CHECKED_OUT",
                            "AVAILABLE"
                        ],
                        "example": "AVAILABLE"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Book"
                }
            },
...
        },
        "securitySchemes": {
            "access_token": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}